+++ gradsflow/data/image.py
@@ -12,59 +12,18 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-import io
 import logging
 import os
 from pathlib import Path
-from typing import Callable, Dict, List, Optional, Union
+from typing import Dict, List, Optional, Union
 
-from PIL import Image
 from torch.utils.data import DataLoader
 from torchvision import transforms as T
 from torchvision.datasets import ImageFolder
 
-from gradsflow.data.ray_dataset import RayDataset
+from gradsflow.data.ray_dataset import RayDataset, RayImageFolder
 
 logger = logging.getLogger("data.image")
-
-
-class RayImageFolder(RayDataset):
-    """Read image datasets
-    ```
-        root/dog/xxx.png
-        root/dog/xxy.png
-        root/dog/[...]/xxz.png
-
-        root/cat/123.png
-        root/cat/nsdf3.png
-        root/cat/[...]/asd932_.png
-    ```
-    """
-
-    def __init__(self, path, transform: Union[Callable, None] = None):
-        super().__init__(path)
-        self.transform = transform
-
-    @staticmethod
-    def file_to_class(files: Union[str, List]):
-        file_list = []
-        if isinstance(files, (tuple, list)):
-            for file in files:
-                file_list.append(file.split("/")[-2])
-            return file_list
-        return files.split("/")[-2]
-
-    def find_classes(self) -> List[str]:
-        files = self.input_files
-        return sorted(list(set(map(self.file_to_class, files))))
-
-    def __iter__(self):
-        for data in self.ds.iter_rows():
-            x = Image.open(io.BytesIO(data[1]))
-            target = self.file_to_class(data[0])
-            if self.transform:
-                x = self.transform(x)
-            yield x, target
 
 
 def get_augmentations(image_size: tuple = (224, 224), auto_augment_policy: bool = True):
